# server/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/users/new":{"post":{"tags":["user-controller"],"operationId":"newUser","parameters":[{"name":"pwd","in":"query","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/restaurants/{id}/services/new":{"post":{"tags":["restaurant-controller"],"operationId":"newTableService","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TableService"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/restaurants/new":{"post":{"tags":["restaurant-controller"],"operationId":"newRestaurant","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Restaurant"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/booking/new":{"post":{"tags":["reservation-controller"],"summary":"Create a new Reservation","operationId":"newReservation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Reservation"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/users/{id}":{"get":{"tags":["user-controller"],"operationId":"getUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/users/{id}/reservations":{"get":{"tags":["user-controller"],"summary":"Return reservations of the user with id","operationId":"getReservations","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/search/paged/byNameAndCountry":{"get":{"tags":["search-controller"],"summary":"Search for restaurants by name and country then return paged content","operationId":"getByNameAndCountry","parameters":[{"name":"name","in":"query","required":true,"schema":{"type":"string"}},{"name":"country","in":"query","required":true,"schema":{"type":"string"}},{"name":"pageNumber","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"sortBy","in":"query","required":false,"schema":{"type":"string","default":"id"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/search/paged/byCountry":{"get":{"tags":["search-controller"],"summary":"Search for restaurants by country and return paged content","operationId":"getByCountry","parameters":[{"name":"country","in":"query","required":true,"schema":{"type":"string"}},{"name":"pageNumber","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"sortBy","in":"query","required":false,"schema":{"type":"string","default":"id"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/restaurants/{id}/services":{"get":{"tags":["restaurant-controller"],"summary":"Return all table service of the restauraunt with id","operationId":"getTableServices","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/restaurants/{id}/reservations/{date}":{"get":{"tags":["restaurant-controller"],"summary":"Return reservations of the restauraunt with id","operationId":"getReservationsByRestaurantAndDate","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"date","in":"path","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/booking/{restaurantId}/services/{date}":{"get":{"tags":["reservation-controller"],"summary":"Return table service available on date for the restauraunt with id","operationId":"getTableServicesByDate","parameters":[{"name":"restaurantId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"date","in":"path","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/booking/availability/{serviceId}/{date}/{time}/{nGuests}":{"get":{"tags":["reservation-controller"],"summary":"Returns availability","operationId":"getAvailability","parameters":[{"name":"serviceId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"date","in":"path","required":true,"schema":{"type":"string","format":"date"}},{"name":"time","in":"path","required":true,"schema":{"$ref":"#/components/schemas/LocalTime"}},{"name":"nGuests","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/users/delete/{id}":{"delete":{"tags":["user-controller"],"operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/restaurants/services/delete/{id}":{"delete":{"tags":["restaurant-controller"],"summary":"Delete a table service","operationId":"deleteTableService","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/restaurants/reservations/delete/{id}":{"delete":{"tags":["restaurant-controller"],"summary":"Reject a reservation","operationId":"rejectReservation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/restaurants/delete/{id}":{"delete":{"tags":["restaurant-controller"],"summary":"Delete restaurant","operationId":"deleteRestaurant","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"User":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"firstName":{"type":"string"},"lastName":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"}}},"LocalTime":{"type":"object","properties":{"hour":{"type":"integer","format":"int32"},"minute":{"type":"integer","format":"int32"},"second":{"type":"integer","format":"int32"},"nano":{"type":"integer","format":"int32"}}},"Restaurant":{"required":["address","city","country","name","privateMail","privatePhone","publicPhone","seatingCapacity"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"country":{"type":"string"},"city":{"type":"string"},"address":{"type":"string"},"privatePhone":{"type":"string"},"publicPhone":{"type":"string"},"privateMail":{"type":"string"},"publicMail":{"type":"string"},"seatingCapacity":{"type":"integer","format":"int32"}}},"TableService":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"serviceName":{"type":"string"},"daysOfWeek":{"uniqueItems":true,"type":"array","items":{"type":"string","enum":["MONDAY","TUESDAY","WEDNESDAY","THURSDAY","FRIDAY","SATURDAY","SUNDAY"]}},"startTime":{"$ref":"#/components/schemas/LocalTime"},"endTime":{"$ref":"#/components/schemas/LocalTime"},"averageMealDuration":{"maximum":180,"minimum":10,"type":"integer","format":"int32"},"restaurant":{"$ref":"#/components/schemas/Restaurant"}}},"Reservation":{"required":["tableService"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"bookingTimestamp":{"type":"string","format":"date-time"},"date":{"type":"string","format":"date"},"startTime":{"$ref":"#/components/schemas/LocalTime"},"notes":{"type":"string"},"user":{"$ref":"#/components/schemas/User"},"restaurant":{"$ref":"#/components/schemas/Restaurant"},"tableService":{"$ref":"#/components/schemas/TableService"},"nguests":{"type":"integer","format":"int32"}}}}}}